requirements.txt
# Add your requirements here like:
# numpy
# scipy>=0.9
Pillow>=2.7.0
